refactor: consolidate blog post selection into blog/posts module
