--- src/blog/filter.ts
@@ -1,11 +1,0 @@
-import type { CollectionEntry } from "astro:content";
-import { SITE } from "@/config";
-
-const postFilter = ({ data }: CollectionEntry<"blog">) => {
-  const isPublishTimePassed =
-    Date.now() >
-    new Date(data.pubDatetime).getTime() - SITE.scheduledPostMargin;
-  return !data.draft && (import.meta.env.DEV || isPublishTimePassed);
-};
-
-export default postFilter;
