--- src/blog/sort.ts
@@ -1,17 +1,0 @@
-import type { CollectionEntry } from "astro:content";
-import postFilter from "./filter";
-
-const getSortedPosts = (posts: CollectionEntry<"blog">[]) =>
-  posts
-    .filter(postFilter)
-    .sort(
-      (a, b) =>
-        Math.floor(
-          new Date(b.data.modDatetime ?? b.data.pubDatetime).getTime() / 1000,
-        ) -
-        Math.floor(
-          new Date(a.data.modDatetime ?? a.data.pubDatetime).getTime() / 1000,
-        ),
-    );
-
-export default getSortedPosts;
