+++ src/pages/archives/index.astro
@@ -5,7 +5,7 @@ import { getCollection } from "astro:content";
 import Main from "@/layouts/Main.astro";
 import Layout from "@/layouts/Layout.astro";
 import Card from "@/components/Card.astro";
-import getPostsByGroupCondition from "@/blog/group";
+import { postFilter } from "@/blog/posts";
 import { SITE } from "@/config";
 
 // Redirect to 404 page if `showArchives` config is false
@@ -13,7 +13,16 @@ if (!SITE.showArchives) {
   return Astro.redirect("/404");
 }
 
-const posts = await getCollection("blog", ({ data }) => !data.draft);
+const posts = await getCollection("blog", postFilter);
+
+const groupBy = <T,>(items: T[], keyOf: (item: T) => string | number) => {
+  const groups: Record<string | number, T[]> = {};
+  for (const item of items) {
+    const key = keyOf(item);
+    (groups[key] ??= []).push(item);
+  }
+  return groups;
+};
 
 const months = [
   "January",
@@ -35,9 +44,7 @@ const months = [
   <Main pageTitle="Archives" pageDesc="All the articles I've archived.">
     {
       Object.entries(
-        getPostsByGroupCondition(posts, post =>
-          post.data.pubDatetime.getFullYear()
-        )
+        groupBy(posts, post => post.data.pubDatetime.getFullYear())
       )
         .sort(([yearA], [yearB]) => Number(yearB) - Number(yearA))
         .map(([year, yearGroup]) => (
@@ -45,10 +52,7 @@ const months = [
             <span class="text-2xl font-bold">{year}</span>
             <sup class="text-sm">{yearGroup.length}</sup>
             {Object.entries(
-              getPostsByGroupCondition(
-                yearGroup,
-                post => post.data.pubDatetime.getMonth() + 1
-              )
+              groupBy(yearGroup, post => post.data.pubDatetime.getMonth() + 1)
             )
               .sort(([monthA], [monthB]) => Number(monthB) - Number(monthA))
               .map(([month, monthGroup]) => (
