+++ src/pages/index.astro
@@ -3,7 +3,7 @@ import { getCollection } from "astro:content";
 import Layout from "@/layouts/Layout.astro";
 import LinkButton from "@/components/LinkButton.astro";
 import Card from "@/components/Card.astro";
-import getSortedPosts from "@/blog/sort";
+import { getSortedPosts } from "@/blog/posts";
 
 const posts = await getCollection("blog");
