+++ src/pages/rss.xml.ts
@@ -3,7 +3,7 @@ export const prerender = true;
 import rss from "@astrojs/rss";
 import { getCollection } from "astro:content";
 import { getPath } from "@/blog/path";
-import getSortedPosts from "@/blog/sort";
+import { getSortedPosts } from "@/blog/posts";
 import { SITE } from "@/config";
 
 export async function GET() {
